Flywheel prototype first adding PID
Tried implementing PID using the built in PID capabilities from RobotC. Got it working at some point, except we didn't know how to get the encoder values. In trying to figure out how to get the encoder values, we managed to break the PID that we had, but we did figure out how to read the encoder values. Next steps: fix the PID control.